Add node 0.10 support
path.isAbsolute was the only entity that had required node 0.11 and
greater node version. Removed that to use the old trick to find if a
path is absolute. Although there were modules that did this, this trick would just
work both on posix as well as windows. When everyone starts moving to
node 0.12, we ll go back to path.isAbsolute.

Keeping npm at version 3, since we believe deduping is the right way to do modules
in node, and flat tree is better.

Tests are now run on node 0.10 on travis too